services/ocr-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.4
pillow==10.2.0
httpx==0.27.0
pdf2image==1.17.0
python-multipart==0.0.9
pytesseract==0.3.10
anthropic==0.21.3
